feat: move metric card size to an enum [MA-2139]
