+++ packages/analytics/metric-cards/src/components/MetricCardContainer.vue
@@ -48,7 +48,7 @@
 import type { PropType } from 'vue'
 import { computed } from 'vue'
 import approxNum from 'approximate-number'
-import { MetricCardSize } from '../constants'
+import { MetricCardSize } from '../enums'
 import type { MetricCardDef, MetricCardDisplayValue } from '../types'
 import { changePolarity, metricChange, defineIcon, calculateChange } from '../utilities'
 import MetricsCard from './display/MetricsCard.vue'
