+++ packages/analytics/metric-cards/src/enums/index.ts
@@ -1,1 +1,2 @@
+export * from './metric-card-size.enum'
 export * from './metric-card-type.enum'
