+++ packages/analytics/metric-cards/src/index.ts
@@ -3,5 +3,4 @@ export { MetricCardContainer }
 
 export * from './types'
 export * from './enums'
-export * from './constants'
 export * from './utilities'
